Uploaded code to Create and
Primary functionality is in GenericSaveLoad.cs, gotten from https://www.riccardostecca.net/articles/save_and_load_mesh_data_in_unity/
A GameObject with the MeshSaver script will take a mesh from the MeshFilter component and write an uncompressed mesh binary. Since the assets we got already come with mesh assets, we may only need to use this if they can't be read by MeshLoader.
A GameObject with the MeshLoader script will read the data from a raw mesh data file to build a mesh that will be added to the MeshFilter. May not be compatible with the pre-existing mesh assets already bundled with the prop assets we got.