Made First Prop Movement Scripts
zip contains the test project, including the input asset I used. Updated the names of the saver and loader to reflect their automatic nature. Also includes SerializableMeshInfo because I uploaded to the wrong branch.
WhatClickOn goes on the player object and casts a ray to get an object within the scene. Double clicks, which is countered with a bool, but it may break other things. Works in tandem with the MovingProp script, which determines the logic of how the prop is changed with the movement of the mouse. Default movement is XY repositioning.