Deleting, Adding, Constraining, Exclusivity
Started implementing Scriptable Objects for storing values that we may want to change. They rely on the GenericNumVar script since they inherit it, and only floats and Vector3's are implemented.
Updated MeshLoader to work with CreateProp to allow for manual additions of props to the scene. Attach the CreateProp script to some object with the PropPreFab attached to it, attach that object to a button's OnClick (like the searchable buttons), and use the string overload of the CreateTheProp() function. It can create a prop in this specific context. Also, MeshLoaderAuto is basically obsolete now outside of testing the internet file loading without dealing with too much code variables.
Added deleting props to the MovingProp script. Additionally, added PropTracker, which attaches to props to keep them in bounds based on a Vector3Var and 3 FloatVars. Those Var assets are included in the zip. Also, PropTracker should theoretically prevent other users from clicking on a prop while it's in use, based on two booleans, which depends on the updated MovingProp and WhatClickOn scripts, but I haven't been able to check because we haven't implemented Multi-User stuff yet.