9th build, URI decoding
To make sure the parameters aren't %XY escaped like in URLs of higher stature. Also, made the final message sent a "Leave" to make the server tell other clients to forget the sending (and leaving) user.
Hopefully this will be the last, with the most related functionality edits being making the server send users in room along with the props.