10th build, testing building
To save hassle, I've been copy-pasting the entire index file into the build after it's built. It could possibly be the problem. If the index file has a version number in its script tag but still doesn't have correct URI decoding, then the index file (or code) may be to blame. Otherwise, it could be error in terms of uploading to server.